requirements.txt
openai-whisper
numpy
scipy
sounddevice
colorama
pyttsx3